Use forgiving casts in msql_srv
usize/isize are treated less aggressively, and allow us to work around
limitations of `ValueInner`'s type granularity.

+++ src/soup_backend.rs
@@ -539,8 +539,8 @@ impl SoupBackend {
         let write_column = |rw: &mut RowWriter<W>, c: DataType| {
             let written = match c {
                 DataType::None => rw.write_col(None::<i32>),
-                DataType::Int(i) => rw.write_col(i as i32),
-                DataType::BigInt(i) => rw.write_col(i as i32),
+                DataType::Int(i) => rw.write_col(i as isize),
+                DataType::BigInt(i) => rw.write_col(i as isize),
                 DataType::Text(t) => rw.write_col(t.to_str().unwrap()),
                 dt @ DataType::TinyText(_) => rw.write_col(dt.to_string()),
                 dt @ DataType::Real(_, _) => {
